src/python/transcription/requirements.txt
# Whisper Transcription Requirements
# Install with: pip install -r requirements.txt

# Core Whisper dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
ffmpeg-python>=0.2.0

# Standard library dependencies (usually included with Python)
# These are listed for completeness
pathlib2>=2.3.7; python_version<'3.4'